rewritten automake/autoconf scripts
fftw.m4 and progs.m4 macros are left unchanged,
the rest is largely rewritten.

removed flags: --with-clipper --with-mmdb --with-ccp4
use CPPFLAGS=-I... and LDFLAGS=-L... instead, if needed

To work with our ccp4/configure script, when library is not found
only a warning is given. This may be changed in the future.

use build-aux directory as AC_CONFIG_AUX_DIR, this name is proposed
in a few places, e.g.:
http://lists.gnu.org/archive/html/bug-gnulib/2005-03/msg00161.html

Use m4 directory for m4 macros, it's the most popular name for this directory

Non-recursive automake is now used, so Makefile generated by autoconf
will be only in the top directory and should not collide with Kevin's
Makefiles.